module hierarchy set
